Streamline AmplitudeCache::reset() bottleneck, first try.
During the read step for builtin Matchbox processes, AmplitudeCache::reset()
is the major contributor to profiling, with ~5% of runtime spent there.
This is a first minimal attempt at streamlining it. If needed, the next step is to take the bool flag out of the maps into a parallel second map structure.